Fixed h2 null
There was an issue with the h2 web console showing a "nul" database schema.
The problem was that we have to use the exact same path of the JDBC URL and the db.url in the application.properties file.
In the .prop file, we had set: jdbc:h2:mem:springrecipes
while in the web console tried to set JDBC URL to:
jdbc:h2:tcp://localhost/C:/Projects/spring-recipes/data/springrecipes